more modular tests
