+++ tests/bandits/test_epsilon_greedy_bandit.py
