+++ tests/bandits/test_softmax_bandit.py
