+++ tests/bandits/test_ucb1_bandit.py
