+++ tests/bandits/test_update_average_rewards_per_arm.py
